Fix initialize process
