+++ change_resistance/Pin_Assignments.h
@@ -1,0 +1,20 @@
+#pragma once
+
+// rotary encoder
+//// pin difinition
+#define pin_a 2
+#define pin_b 3
+//// define rotate directions
+#define d_still 0
+#define d_right 1
+#define d_left  2
+
+// digital potentiometer
+//// pin difinition
+#define pin_monitor_r 0
+//// parameters
+#define pin_spi_sck  13
+#define pin_spi_miso 12
+#define pin_spi_mosi 11
+#define pin_spi_nCS1 10
+#define pin_spi_nCS2 8
