+++ change_resistance/change_resistance.ino
@@ -45,28 +45,10 @@
 
 #include <Wire.h>
 #include <SPI.h>
-#include "potentiometer.h"
+#include "potentiometer.hpp"
 
 
 #define debug_mode 1
-// rotary encoder
-//// pin difinition
-#define pin_a 2
-#define pin_b 3
-//// define rotate directions
-#define d_still 0
-#define d_right 1
-#define d_left  2
-
-// digital potentiometer
-//// pin difinition
-#define pin_monitor_r 0
-//// parameters
-#define pin_spi_sck  13
-#define pin_spi_miso 12
-#define pin_spi_mosi 11
-#define pin_spi_nCS1 10
-#define pin_spi_nCS2 8
 
 // variables for rotary_encoder
 volatile int a = 0;          // A相の値
@@ -81,6 +63,7 @@ int num_potentiometer = 0;
 
 
 void setup() {
+  setup_potentiometer();
   setup_rotary_encoder();
   setup_digital_potentiometer();
   Serial.begin(9600);
